requirements.txt
numpy==1.24.4
opencv-python-headless==4.10.0.84
pillow==10.4.0
streamlit==1.38.0